# Optimise some things a bit