A couple of optimisations:
- Index the string cache by pointer instead of value, since it's quicker to hash and compare, and we're optimising for caching the exact same literal, which will always have the same pointer.
- Use a reference to the strings in the string cache instead of cloning, to avoid some overhead.
- Specialize `deserialize_seq` to index into arrays instead of iterating. (Note: the bigger effect here is probably the fact that indexing is infallible, which avoids the overhead of `try ... catch`.)
- Specialize `deserialize_map` to index into `Object.entries` instead of iterating.
- Serialize arrays by setting indices instead of using `push`. (No idea why this is faster, but it is.)

| Benchmark              | Before        | After         | Change  |
| ---------------------- | ------------- | ------------- | ------- |
| Parse canada           | 46.56 ms/iter | 25.71 ms/iter | -44.78% |
| Parse citm_catalog     | 10.39 ms/iter | 7.12 ms/iter  | -31.47% |
| Parse twitter          | 9.11 ms/iter  | 8.08 ms/iter  | -11.31% |
| Serialize canada       | 11.84 ms/iter | 8.88 ms/iter  | -25.00% |
| Serialize citm_catalog | 7.62 ms/iter  | 6.23 ms/iter  | -18.24% |
| Serialize twitter      | 14.99 ms/iter | 14.28 ms/iter | -4.74%  |