@RReverser 2022-05-04
Thanks. I think I had a slightly simpler optimisation branch for arrays somewhere, let me dig that up first before looking deeper.