@RReverser 2022-05-06
> * Specialize `deserialize_seq` to index into arrays instead of iterating. (Note: the bigger effect here is probably the fact that indexing is infallible, which avoids the overhead of `try ... catch`.)
> * Specialize `deserialize_map` to index into `Object.entries` instead of iterating.

Looking at my branch, I did these optimizations too, but instead of extending my own `*Access` types, I used `serde::de::value::{Seq, Map}Deserializer`. That way, there's no need to go via wrapping-unwrapping `Result`s for the "source is an array" case.

I have some more questions about other optimizations, but if you could change the PR to use those serde built-ins too, we can probably avoid some merge conflicts + make the code a bit simpler.