@Liamolucko 2022-05-07
> I have some more questions about other optimizations, but if you could change the PR to use those serde built-ins too, we can probably avoid some merge conflicts + make the code a bit simpler.

Okay, done.