@RReverser 2022-05-11
Can you rebase to clean up the history a bit? I could squash everything, but I'd prefer to keep the optimisations in separate commits like you did, just drop the commits with reverted changes (and reverts themselves).