@RReverser 2022-05-12
Added one more [small optimisation](https://github.com/cloudflare/serde-wasm-bindgen/commit/68c5be4c42a6c1cba360259be1db36fad1cd660b) and released as 0.4.3.